+++ plugins/qcheck-stm/src/config.ml
@@ -2,6 +2,24 @@ open Gospel
 open Ortac_core
 open Ppxlib
 
+module FrequenciesMap = struct
+  module M = Map.Make (String)
+
+  type t = (int * location) M.t
+
+  let empty = M.empty
+
+  let add label_loc v =
+    let k = label_loc.txt and loc = label_loc.loc in
+    M.add k (v, loc)
+
+  let pop k m =
+    let res = Option.value ~default:1 @@ Option.map fst @@ M.find_opt k m in
+    (res, M.remove k m)
+
+  let unused m = List.map (fun (k, (_freq, loc)) -> (k, loc)) @@ M.bindings m
+end
+
 type config_under_construction = {
   sut_core_type' : Ppxlib.core_type option;
   init_sut' : Ppxlib.expression option;
@@ -9,6 +27,7 @@ type config_under_construction = {
   pp_mod' : Ppxlib.structure option;
   ty_mod' : Ppxlib.structure option;
   cleanup' : Ppxlib.structure_item option;
+  frequencies' : FrequenciesMap.t;
 }
 
 let config_under_construction =
@@ -18,6 +37,7 @@ let config_under_construction =
     gen_mod' = None;
     pp_mod' = None;
     ty_mod' = None;
+    frequencies' = FrequenciesMap.empty;
     cleanup' = None;
   }
 
@@ -30,6 +50,7 @@ type t = {
   pp_mod : Ppxlib.structure option; (* Containing custom pretty printers *)
   ty_mod : Ppxlib.structure option; (* Containing custom STM.ty extensions *)
   cleanup : Ppxlib.structure_item option;
+  frequencies : FrequenciesMap.t;
   module_prefix : string option;
   submodule : string option;
   domain : bool;
@@ -51,6 +72,7 @@ let mk_config context module_prefix submodule domain count cfg_uc =
   and gen_mod = cfg_uc.gen_mod'
   and pp_mod = cfg_uc.pp_mod'
   and ty_mod = cfg_uc.ty_mod'
+  and frequencies = cfg_uc.frequencies'
   and cleanup = cfg_uc.cleanup' in
   ok
     {
@@ -61,6 +83,7 @@ let mk_config context module_prefix submodule domain count cfg_uc =
       gen_mod;
       pp_mod;
       ty_mod;
+      frequencies;
       cleanup;
       module_prefix;
       submodule;
@@ -191,6 +214,28 @@ let module_binding cfg_uc (mb : Ppxlib.module_binding) =
   | Some name when String.equal "Ty" name ->
       let* content = get_structure name mb in
       ok { cfg_uc with ty_mod' = Some content }
+  | Some name when String.equal "Frequencies" name ->
+      let* content = get_structure name mb in
+      let open Ast_pattern in
+      let destruct =
+        pstr_value drop
+          (value_binding ~pat:(ppat_var __') ~expr:(eint __) ^:: nil)
+      and error stri =
+        let loc = stri.pstr_loc in
+        error (Ill_formed_frequency, loc)
+      in
+      let aux acc stri =
+        parse destruct Location.none
+          ~on_error:(Fun.const @@ error stri)
+          stri
+          (fun name freq ->
+            ok
+              {
+                acc with
+                frequencies' = FrequenciesMap.add name freq acc.frequencies';
+              })
+      in
+      fold_left aux cfg_uc content
   | _ -> ok cfg_uc
 
 let scan_config cfg_uc config_mod =