+++ plugins/qcheck-stm/test/all_warnings_config.ml
@@ -3,3 +3,9 @@ open All_warnings
 let init_sut = make 16 'a'
 
 type sut = char t
+
+module Frequencies = struct
+  let bob = 42
+  let lenght = bob
+  let constant = 42
+end