+++ plugins/qcheck-stm/test/stack_stm_tests.expected.ml
@@ -82,16 +82,16 @@ module Spec =
       let open QCheck in
         make ~print:show_cmd
           (let open Gen in
-             oneof
-               [(pure (fun () -> Create ())) <*> unit;
-               (pure (fun v -> Push v)) <*> char;
-               pure Pop;
-               pure Pop_opt;
-               pure Top;
-               pure Top_opt;
-               pure Clear;
-               pure Copy;
-               pure Is_empty])
+             frequency
+               [(1, ((pure (fun () -> Create ())) <*> unit));
+               (1, ((pure (fun v -> Push v)) <*> char));
+               (1, (pure Pop));
+               (1, (pure Pop_opt));
+               (1, (pure Top));
+               (1, (pure Top_opt));
+               (1, (pure Clear));
+               (1, (pure Copy));
+               (1, (pure Is_empty))])
     let next_state cmd__002_ state__003_ =
       match cmd__002_ with
       | Create () ->