+++ plugins/qcheck-stm/test/submodule_tests.expected.ml
@@ -62,7 +62,8 @@ module Spec =
       let open QCheck in
         make ~print:show_cmd
           (let open Gen in
-             oneof [(pure (fun a_1 -> Make a_1)) <*> small_signed_int])
+             frequency
+               [(1, ((pure (fun a_1 -> Make a_1)) <*> small_signed_int))])
     let next_state cmd__002_ state__003_ =
       match cmd__002_ with
       | Make a_1 ->